Move react dom render out of the component file
